requirements.txt
flask==2.2.3
aiogram>=3.0.0
sqlalchemy==2.0.7
passlib==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.0
Werkzeug==2.2.3
nest_asyncio==1.5.7
requests==2.31.0
